local_frontend/requirements.txt
streamlit==1.48.1
requests==2.32.0
pillow==10.4.0